+++ examples/nuxt-express-session/routes/>0-init.js
@@ -1,6 +1,0 @@
-export default (req, res, next) => {
-  req.request = `${req.method} ${req.path}`
-  req.isSigned = (req.session.user !== undefined)
-
-  next()
-}