+++ examples/nuxt-express-session/routes/auth/>redirect-if-signed.js
@@ -1,7 +1,0 @@
-export default (req, res, next) => {
-  if (req.isSigned && req.method === 'GET') {
-    return res.redirect('/')
-  }
-
-  next()
-}