+++ examples/restful-lowdb/api/posts/get.js
@@ -1,5 +1,0 @@
-import Post from '~model/post'
-
-export default (req, res) => {
-  res.json(Post.newest(100))
-}