+++ test/index.test.js
@@ -1,6 +1,0 @@
-import test from 'ava'
-import dynapi from '../index.js'
-
-test('Dynapi.js Class', t => {
-  t.is(typeof dynapi, 'function')
-})